+++ src/View/Helper/AbstractMenuHelper.php
@@ -11,7 +11,6 @@ declare(strict_types=1);
  * @copyright   Copyright (c) Mirko Pagliai
  * @link        https://github.com/mirko-pagliai/me-cms
  * @license     https://opensource.org/licenses/mit-license.php MIT License
- * @since       2.31.9
  */
 
 namespace MeCms\View\Helper;
@@ -19,8 +18,9 @@ namespace MeCms\View\Helper;
 use Cake\View\Helper;
 
 /**
- * `AbstractMenuHelper` for all `MenuHelper` classes
- * @see \MeCms\View\Helper\MenuBuilderHelper::generate() for more information
+ * AbstractMenuHelper.
+ *
+ * Other "menu helpers" have to extend this abstract class and implement its missing methods.
  * @property \MeCms\View\Helper\IdentityHelper $Identity
  */
 abstract class AbstractMenuHelper extends Helper
@@ -30,4 +30,43 @@ abstract class AbstractMenuHelper extends Helper
      * @var array
      */
     public $helpers = ['MeCms.Identity'];
+
+    /**
+     * @var string
+     */
+    protected string $name;
+
+    /**
+     * Gets the links for this menu. Each links is an array of parameters
+     * @return array[]
+     */
+    abstract public function getLinks(): array;
+
+    /**
+     * Gets the name of the current plugin
+     * @return string
+     * @throws \ReflectionException
+     */
+    public function getName(): string
+    {
+        if (empty($this->name)) {
+            $plugin = substr(get_class($this), 0, strpos(get_class($this), '\View\Helper') ?: 0);
+
+            $this->name = strtolower(str_replace('\\', '-', $plugin) . '-' . substr(get_class_short_name($this), 0, -10));
+        }
+
+        return $this->name;
+    }
+
+    /**
+     * Gets the options for this menu
+     * @return array
+     */
+    abstract public function getOptions(): array;
+
+    /**
+     * Gets the title for this menu
+     * @return string
+     */
+    abstract public function getTitle(): string;
 }