+++ src/View/Helper/WidgetHelper.php
@@ -42,8 +42,15 @@ class WidgetHelper extends Helper
                 [$name, $args] = [array_key_first($args), array_value_first($args)];
             }
 
-            /** @var array-key $name */
-            return is_int($name) && is_string($args) ? [$args => []] : [$name => $args];
+            if (is_int($name) && is_string($args)) {
+                return [$args => []];
+            }
+
+            /**
+             * @var string $name
+             * @var array $args
+             */
+            return [$name => $args];
         })->toList() : [];
     }