+++ templates/element/admin/sidebar.php
@@ -1,4 +1,5 @@
 <?php
+/** @noinspection PhpUnhandledExceptionInspection */
 declare(strict_types=1);
 
 /**
@@ -14,38 +15,32 @@ declare(strict_types=1);
  *
  * @var \MeCms\View\View\Admin\AppView $this
  */
-
-use Cake\Utility\Inflector;
-use MeCms\Core\Plugin;
-
-$pluginsMenus = array_merge(...array_map(fn(string $plugin): array => $this->MenuBuilder->generate($plugin), Plugin::all(['mecms_core' => false])));
-$pluginsNames = array_map(fn($name): string => Inflector::camelize(str_replace(['/', '.'], '_', $name)), array_keys($pluginsMenus));
 ?>
 
 <div id="accordionSidebar" class="accordion accordion-flush">
-    <?php foreach (array_combine($pluginsNames, $pluginsMenus) as $name => $menu) : ?>
+    <?php foreach ($this->getAllMenuHelpers() as $Helper) : ?>
     <div class="accordion-item">
-        <h2 class="accordion-header" id="heading<?= $name ?>">
+        <h2 class="accordion-header" id="heading-<?= $Helper->getName() ?>">
             <?php
-            $titleOptions = optionsParser($menu['titleOptions'])->add([
+            $options = optionsParser($Helper->getOptions())->add([
                 'class' => 'accordion-button collapsed p-3',
                 'data-bs-toggle' => 'collapse',
-                'data-bs-target' => '#collapse' . $name,
+                'data-bs-target' => '#collapse-' . $Helper->getName(),
                 'aria-expanded' => 'false',
-                'aria-controls' => 'collapse' . $name,
+                'aria-controls' => 'collapse-' . $Helper->getName(),
             ]);
 
-            echo $this->Html->link($menu['title'], '#collapse' . $name, $titleOptions->toArray());
+            echo $this->Html->link($Helper->getTitle(), '#collapse-' . $Helper->getName(), $options->toArray());
             ?>
         </h2>
-        <div id="collapse<?= $name ?>" class="accordion-collapse collapse" aria-labelledby="heading<?= $name ?>" data-bs-parent="#accordionSidebar">
+        <div id="collapse-<?= $Helper->getName() ?>" class="accordion-collapse collapse" aria-labelledby="heading-<?= $Helper->getName() ?>" data-bs-parent="#accordionSidebar">
             <div class="accordion-body my-2 p-0">
                 <ul class="list-group list-group-flush">
-                <?php foreach ($menu['links'] as $link) : ?>
-                    <li class="list-group-item p-0">
-                        <?= $this->Html->link($link[0], $link[1], ['class' => 'd-block px-3 py-2']) ?>
-                    </li>
-                <?php endforeach; ?>
+                    <?php foreach ($Helper->getLinks() as $link) : ?>
+                        <li class="list-group-item p-0">
+                            <?= $this->Html->link($link[0], $link[1], ['class' => 'd-block px-3 py-2']) ?>
+                        </li>
+                    <?php endforeach; ?>
                 </ul>
             </div>
         </div>