+++ CHANGELOG.md
@@ -1,4 +1,15 @@
 # 2.x branch
+## 2.32 branch
+### 2.32.0
+* as for `me-tools` 2.24.0, it uses `Configure::readFromPlugins()` with `ConfigFiles`, `WritableDirs` and `VendorLinks`
+  keys instead of the old `CONFIG_FILES`, `WRITABLE_DIRS` and `VENDOR_LINKS` keys;
+* added `MeCms\View\View\Admin\AppView::getAllMenuHelpers()` method. The admin sidebar template has been rewritten;
+* the `MenuHelper` has been removed and replaced with more classes under the `MeCms\View\Helper\MenuHelper` namespace
+  (one class for each menu, instead of one method for each menu in the helper itself). The helper list is set in the
+  bootstrap, with the `MeCms.MenuHelpers` config key (and the other plugins will have to do the same);
+* `AbstractMenuHelper` has been completely rewritten, with new methods to be implemented by menu helpers that extend it;
+* `MenuBuilderHelper` no longer exists (and is not being replaced).
+
 ## 2.31 branch
 ### 2.31.9
 * fixed little bug for `IdentityHelper::isGroup()`;